requirements.txt
openai-whisper==20240930
PyQt6==6.8.1
